#6706 split table component into two smaller entities
remove unused code

# Conflicts:
#	frontend/src/utils/tableHelpers.js
