+++ frontend/src/containers/Table.js
@@ -16,7 +16,7 @@ import { openModal } from '../actions/WindowActions';
 import { containerPropTypes } from '../utils/tableHelpers';
 import { mapIncluded } from '../utils/documentListHelper';
 
-import Table from '../components/table/Table';
+import Table from '../components/table/TableWrapper';
 
 class TableContainer extends PureComponent {
   componentWillUnmount() {
@@ -252,7 +252,6 @@ const mapStateToProps = (state, props) => {
     rows: table.rows,
     columns: table.columns,
     selected: table.selected,
-    pending: table.dataPending,
     collapsedParentRows: table.collapsedParentRows,
     collapsedRows: table.collapsedRows,
     collapsedArrayMap: table.collapsedArrayMap,
